extract load ts config to util
